requirements.txt
certifi==2024.7.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.3.2
idna==3.7
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.4
MarkupSafe==2.1.1
pip==23.3
python-dotenv==0.21.0
requests==2.32.0
setuptools==70.0.0
shortuuid==1.0.11
urllib3==1.26.19
Werkzeug==3.0.3
wheel==0.38.4
zipp==3.19.1
flask-login==0.6.3
geopy==2.3.0